[Change] serial2yyyymmdd: enforces uint32 on out

+++ utils/serial2yyyymmdd.m
@@ -1,4 +1,4 @@
 function x = serial2yyyymmdd(n)
 x = datevec(n);
-x = x(:,1:3)*[10000; 100; 1];
+x = uint32(x(:,1:3)*[10000; 100; 1]);
 end
